requirements.txt
openai
langchain
langchain-openai
langchain-core
langchain-community
python-dotenv
